+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/udt/UDTIndexingAT.java
@@ -18,6 +18,7 @@
 
 package com.stratio.cassandra.lucene.testsAT.udt;
 
+import com.datastax.driver.core.SimpleStatement;
 import com.datastax.driver.core.exceptions.DriverException;
 import com.stratio.cassandra.lucene.testsAT.BaseAT;
 import com.stratio.cassandra.lucene.testsAT.util.CassandraUtils;
@@ -500,4 +501,24 @@ public class UDTIndexingAT extends BaseAT {
         assertTrue("Selecting a non-existent type inside udt inside udt must return an Exception", true);
     }
 
+    @Test
+    public void testNonCompleteUDT() {
+
+        String insert = "INSERT INTO " +
+                        cassandraUtils.getKeyspace() +
+                        "." +
+                        cassandraUtils.getTable() +
+                        "(login, first_name, last_name, address) VALUES (" +
+                        "'USER10'," +
+                        "'Tom'," +
+                        "'Smith',{"+
+                        "city: 'Madrid'});";
+
+        cassandraUtils.execute(new SimpleStatement(insert));
+        cassandraUtils.refresh();
+
+        CassandraUtilsSelect select = cassandraUtils.filter(match("address.city","Madrid"));
+        assertEqualsAndOnlyThisString(select.stringColumn("login"),
+                                      new String[]{"USER10"});
+    }
 }