+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 * Returns static columns (#70)
 * Fixed UDT bug (#85)
+* Sort by geological distance
     
 ## 2.2.4.0 (11 December 2015)