+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.2.4.1 (12 January 2016)
+
+* Returns static columns (#70)
+* Fixed UDT bug (#85)
+* Sort by geological distance
+
 ## 2.2.4.0 (11 December 2015)
 
 * Upgrade to Apache Cassandra 2.2.4